Implements bundler parser
